fix(editor): fix bracket keymap order; rename json-editor to editor
